backend/requirements.txt
# ============== backend/requirements.txt ==============
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.3.0
pandas==2.1.3
numpy==1.24.3
clickhouse-connect==0.6.22
econml==0.14.1
scikit-learn==1.3.2
prophet==1.1.5
plotly==5.18.0
holidays==0.37
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6